rename radiator.js to radiator.js
